refactor: test file 디렉토리 분리
