+++ src/test/java/roomescape/service/reservation/ReservationThemeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.service.reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -26,9 +26,6 @@ import roomescape.dto.response.ReservationTimeResponse;
 import roomescape.global.exception.roomescape.RoomEscapeErrorStatus;
 import roomescape.global.exception.roomescape.RoomEscapeException;
 import roomescape.service.member.MemberService;
-import roomescape.service.reservation.ReservationService;
-import roomescape.service.reservation.ReservationThemeService;
-import roomescape.service.reservation.ReservationTimeService;
 
 @ActiveProfiles("test")
 @SpringBootTest
